refactor: use strategy pattern for handling of shared incoming changes
