+++ src/helpers/strategies/ChangeStrategy.ts
@@ -1,0 +1,5 @@
+import { Payload } from "../InboxProcessor";
+
+export interface ChangeStrategy {
+  execute(payload: Payload, relId: string): Promise<void>;
+}
